upgrades: use dynamic routes not file based routes
